+++ RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+v2.2.1
+--------
+
+* Make code for converting comma-delimited strings to array more robust.
+  Now, extra spaces are much less likely to confuse it.
+
 v2.2.0
 ------